# Conflicts:
#	src/main/kotlin/com/sparta/techTree/like/service/LikeServiceImpl.kt
#	src/main/resources/application.yml

+++ src/main/kotlin/com/sparta/techTree/post/service/PostServiceImpl.kt
@@ -10,14 +10,21 @@ import com.sparta.techTree.post.model.Post
 import com.sparta.techTree.post.model.toResponse
 import com.sparta.techTree.post.repository.PostRepository
 import org.springframework.data.repository.findByIdOrNull
-
 import com.sparta.techTree.like.repository.LikeRepository
+import com.sparta.techTree.user.model.UserEntity
+import com.sparta.techTree.user.repository.UserRepository
+import java.nio.file.AccessDeniedException
 
 
 @Service
 
-class PostServiceImpl(private val postRepository: PostRepository, private val likeRepository: LikeRepository) :
+class PostServiceImpl(
+    private val postRepository: PostRepository,
+    private val likeRepository: LikeRepository,
+    private val userRepository: UserRepository
+) :
     PostService {
+
 
     override fun getPostList(): List<PostResponse> {
         return postRepository.findAll().map { post ->
@@ -36,29 +43,41 @@ class PostServiceImpl(private val postRepository: PostRepository, private val li
     }
 
 
-    @Transactional
-    override fun createPost(request: CreatePostRequest): PostResponse {
+    override fun createPost(request: CreatePostRequest, userId: Long): PostResponse {
+        val user: UserEntity = userRepository.findByIdOrNull(userId) ?: throw ModelNotFoundException("User", userId)
         val createdPost = postRepository.save(
             Post(
-                title = request.title, content = request.content, userId = request.userId, countLikes = 0
+                title = request.title,
+                content = request.content,
+                user = user,
+                countLikes = 0
             )
         )
         return createdPost.toResponse()
     }
 
     @Transactional
-    override fun updatePost(postId: Long, request: UpdatePostRequest): PostResponse {
+    override fun updatePost(postId: Long, userId: Long, request: UpdatePostRequest): PostResponse {
         val post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
-        val countLikes = likeRepository.countByPostId(postId)
+        if (post.user.id != userId) {
+            throw AccessDeniedException("User with ID $userId does not have permission to update post with ID $postId")
+        }
         post.title = request.title ?: post.title
         post.content = request.content ?: post.content
+        val updatePost = postRepository.save(post)
+        val countLikes = likeRepository.countByPostId(postId)
         post.countLikes = countLikes
-        return post.toResponse()
+
+
+        return updatePost.toResponse()
     }
 
     @Transactional
-    override fun deletePost(postId: Long) {
+    override fun deletePost(postId: Long, userId: Long) {
         val post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
+        if (post.user.id != userId) {
+            throw AccessDeniedException("User with ID $userId does not have permission to update post with ID $postId")
+        }
         postRepository.delete(post)
     }
 }
